server/workers/persistence/requirements.txt
flask
flask-cors
flask_sqlalchemy
sqlalchemy_utils
flask_restx
Werkzeug
gunicorn
psycopg2-binary
